Prevent un-clicked elements from receiving focus
When I click on a checkbox, radio, or button in a cell with multiple such inputs, the last element in the cell is focused. This is confusing for the user. If we just focus `input[type="text"]` on a click in a cell, we prevent this problem.

See demo: https://www.youtube.com/watch?v=7D0eZJ2V41M&feature=youtu.be