Override styles when printing to ensure good output.
After **numerous** trials and errors with the help from @kumailht we finally were able to override some settings when printing GridForms to ensure the forms look as good on paper as they do on the screen.

To ensure these styles are activated (i.e. the media type is set to print) use the following in the wkhtmltopdf command line:

--print-media-type

If using `wicked_pdf` to generate PDFs in Rails, you can use `print_media_type: true` configuration option.